pyproject.toml
[tool.poetry]
name = "webp2u"
version = "0.1.0"
description = ""
authors = ["Dzianis Bazhok <undying.k@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
Pillow-PIL = "^0.1dev"
python-telegram-bot = "^13.7"
ffmpeg-python = "^0.2.0"

[poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
